Reuse core component inspector view
